requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
async-timeout==4.0.3
attrs==25.4.0
brotli==1.2.0
certifi==2026.2.25
charset-normalizer==3.4.4
click==8.1.8
dataclasses-json==0.6.7
ddgs==9.8.0
duckduckgo_search==8.1.1
easyocr==1.7.2
exceptiongroup==1.3.1
fastapi==0.128.8
filelock==3.19.1
frozenlist==1.8.0
fsspec==2025.10.0
h11==0.16.0
h2==4.3.0
hf-xet==1.3.2
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==0.36.2
hyperframe==6.1.0
idna==3.11
ImageIO==2.37.2
Jinja2==3.1.6
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.27
langchain-community==0.3.31
langchain-core==0.3.83
langchain-text-splitters==0.3.11
langsmith==0.4.37
lazy_loader==0.4
lxml==6.0.2
MarkupSafe==3.0.3
marshmallow==3.26.2
mpmath==1.3.0
multidict==6.7.1
mypy_extensions==1.1.0
mysql-connector-python==9.4.0
networkx==3.2.1
ninja==1.13.0
numpy==2.0.2
opencv-python-headless==4.13.0.92
orjson==3.11.5
packaging==25.0
pillow==11.3.0
primp==0.15.0
propcache==0.4.1
pyclipper==1.3.0.post6
pydantic==2.12.5
pydantic-settings==2.11.0
pydantic_core==2.41.5
PyMuPDF==1.26.5
python-bidi==0.6.7
python-dotenv==1.2.1
PyYAML==6.0.3
regex==2026.1.15
requests==2.32.5
requests-toolbelt==1.0.0
safetensors==0.7.0
scikit-image==0.24.0
scikit-learn==1.6.1
scipy==1.13.1
sentence-transformers==5.1.2
shapely==2.0.7
socksio==1.0.0
SQLAlchemy==2.0.47
starlette==0.49.3
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tifffile==2024.8.30
tokenizers==0.22.2
torch==2.8.0
torchaudio==2.8.0
torchvision==0.23.0
tqdm==4.67.3
transformers==4.57.6
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uuid_utils==0.14.1
uvicorn==0.39.0
yarl==1.22.0
zstandard==0.25.0
